section 5, lesson 40
Now, lets make sure the camera on mission complete moves as it should in every  machine, not only in the one of the locally controller winner/loser as previously.

And thats it, stealth game finished.

To play it with packaged versions, both players need to execute the packaged version

- The server needs to write in console "open FirstPersonExampleMap?listen"  and you need to get your ip and give it to the other players.
- The aditional players needs to write in console "open host.ip.here:7777",  like open 127.0.0.1:7777

Router and firewall restrictions may apply.

Voila